pyproject.toml
[tool.poetry]
name = "tortoise-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.6"}


[tool.poetry.group.dev.dependencies]
yapf = "^0.40.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
